move history item to new file
