+++ src/history/base.rs
@@ -1,9 +1,8 @@
-use std::time::Duration;
-
 use chrono::Utc;
-use serde::{de::DeserializeOwned, Serialize};
 
-use crate::core_editor::LineBuffer;
+use crate::{core_editor::LineBuffer, HistoryItem};
+
+use super::{HistoryItemId, HistorySessionId};
 
 // todo: better error type
 pub type Result<T> = std::result::Result<T, String>;
@@ -26,64 +25,6 @@ pub enum CommandLineSearch {
     Prefix(String),
     Substring(String),
     Exact(String),
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct HistoryItemId(pub(crate) i64);
-impl HistoryItemId {
-    pub(crate) fn new(i: i64) -> HistoryItemId {
-        HistoryItemId(i)
-    }
-}
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct HistorySessionId(pub(crate) i64);
-impl HistorySessionId {
-    pub(crate) fn new(i: i64) -> HistorySessionId {
-        HistorySessionId(i)
-    }
-}
-/// This trait represents additional context to be added to a history (see [HistoryItem])
-pub trait HistoryItemExtraInfo: Serialize + DeserializeOwned + Default + Send {}
-impl HistoryItemExtraInfo for () {}
-/// Represents one run command with some optional historical context
-#[derive(Clone, Debug, PartialEq)]
-pub struct HistoryItem<ExtraInfo: HistoryItemExtraInfo = ()> {
-    /// primary key, unique across one history
-    pub id: Option<HistoryItemId>,
-    /// date-time when this command was started
-    pub start_timestamp: Option<chrono::DateTime<Utc>>,
-    /// the full command line as text
-    pub command_line: String,
-    /// a unique id for one shell session.
-    /// used so the history can be filtered to a single session
-    pub session_id: Option<HistorySessionId>,
-    /// the hostname the commands were run in
-    pub hostname: Option<String>,
-    /// the current working directory
-    pub cwd: Option<String>,
-    /// the duration the command took to complete
-    pub duration: Option<Duration>,
-    /// the exit status of the command
-    pub exit_status: Option<i64>,
-    /// arbitrary additional information that might be interesting
-    pub more_info: Option<ExtraInfo>,
-}
-
-impl HistoryItem {
-    /// create a history item purely from the command line with everything else set to None
-    pub fn from_command_line(cmd: impl Into<String>) -> HistoryItem {
-        HistoryItem {
-            id: None,
-            start_timestamp: None,
-            command_line: cmd.into(),
-            session_id: None,
-            hostname: None,
-            cwd: None,
-            duration: None,
-            exit_status: None,
-            more_info: None,
-        }
-    }
 }
 
 #[derive(Clone, Copy, PartialEq, Eq)]
@@ -231,6 +172,8 @@ mod test {
             more_info: None,
         }
     }
+    use std::time::Duration;
+
     use super::*;
     fn create_filled_example_history() -> Result<Box<dyn History>> {
         #[cfg(feature = "sqlite")]
